Improve CI job efficiency and reporting
This commits aims at improving the CI job mostly in terms of efficency
and in the way coverage reports are generated.

First we limit the types of ROS messages that are generated to Python and
C++. EusLisp, Lisp, Java and Javascript are not used so this should
speed-up the CI job somewhat.

Secondly pytest is now invoqued only once at the root of the repository
and tests relying on Qt QApplication are now excluded by a pytest command
within the file. A some work is still required here to make them work in a
headless environment.

Finally both the codecov and pytest-cov code coverage reports exclude test
scripts in their calculations. This should reflect better code quality.

Signed-off-by: Christophe Paccolat <christophe.paccolat@alumni.epfl.ch>