Added the IMU to the simulator by creating a launch file remapping the correct output.
This merge request is also a first step to implementing a new sandbox-like launch file.
Every satellite has its own group (named after the satellite), which has again a
group for sensors and for actuators. These can then simply be added or removed based on
the needs of the user.

Signed-off-by: Christian Lanegger <clanegge@student.ethz.ch>