This commit holds the changes, which were done to make the IMUnode mergable into the master node. Following changes were done: 	- Created an node integration test for the IMU node and an additional test checking frequency and publish topic. 	- All necessary docstrings in the IMU node have been added. 	- The gitlab ci has been updated to run the IMU-test.
The integration test tests the functionality of the callback of the IMU node's subscriber. In the test it is being
checking if the values returned by the IMU are similar to the published values of the propagator.

The frequency and publish test check if the publish rate is in fact 10hz and if the IMU node
publishes to the correct topic (by convention: /<spacecraft>/imu).

Signed-off-by: Christian Lanegger <clanegge@student.ethz.ch>