Commiting the implementation of the sun sensor model.
The sun sensor model, which was not completely finished, has now been finished.
The model now uses the dat provided in the source mentioned in the docstring.
The plots have been changed to in which the sensor output (Delta_I / I_ref) as well
as its standard deviation is provided unit-less instead of engineering units.

Additionally the proper documentation and unittests have been added. Currently, there
is no node implementing the sun sensor into the API and this will be left as future
work.

Signed-off-by: Christian Lanegger <clanegge@student.ethz.ch>