First working version of new core structure.
-There is now a new top level launch file: simulator.launch
 This file launches all static nodes needed by the simulator.

-Additionally a scenario launch file has been created.
 This loads all scenario parameters (like initial coordinates
 of all spacecrafts) and includes all to be propagated spacecrafts.

-Every spacecraft type has now its own launch file which initializes
 all spacecraft related nodes (sensors, actuators, rviz_tf, ..) and
 loads the sppacecraft related yaml file which holds all configuartions
 of theses initialized nodes as well as propagator settings.

Additionally a parser for the propagator has been created which translates
the parameters from the yaml file so that they can be further processed
in by the Propagator builder.

Another new feature is the ability to load any number of spacecraft into
the simulator. At initialization a list of spacecraft objects is created.
Every spacecraft object has its own propagator, as well as publishers and
subscribers, which are being called at every time-step. This ensures that
there is no mix-up between different propagator output and that the correct
messages are being published to the correct topic.

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>