Made IMU node sensor agnostic by dynamically loading the module
The IMU node was not sensor agnostic because the name of the sensor module and class were hard coded in the node. To make the node more flexible, the name of the sensors module and class are now contained in the spacecraft configuration file and used while initializing the imu node to load the appropriate module and class.

Signed-off-by: Bastien Gorret <bastien.gorret@epfl.ch>