Modified the dzhanibekov simulation so that it runs with new framework.
A T_handle spacecraft has been created which is used in the dzhanibekov
scenario and which also loads the beatiful T-handle mesh
created Michael.

Additionally a small bug in the satellite_tf launch file has been fixed
where a remapping was missing and therefore no lvlh frame was being
published.

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>