Deleted old launch files and adapted test to new framework.
The test are now all launched using the top-level launch file
and therefore testing also the integration of the launch file
structure. A test_mission scenario and a testcraft was created.
The testcraft is completely empty (no sensors, acutators, etc.).
Theses elements are being added in the corresponding test integration
files.

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>