Added features to epoch class to bring it up to date.
The epoch class has two new methods Rate() and time_step(),
which are supposed to be used by all simulator nodes to
run at the same frequency/synchronously with the driving node
as well as to get the simulation time step of the simulator.

The epoch class stores theses values as rosparam and also has two
functions to set those values. These are restricted however to the
SimTimePublisher class.

Also a skip to problematic test in magnetorquer integration tests
was added. The test will be recreated later on in a propagator
integration test on a new branch.

Commit also contains a minor bug fix where the testcraft launch
file was loading the T_handle configuration instead of the one
of the testcraft.

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>