Fixed small bug in DipoleModel test and added EnviSat mission to scenarios.
The bug in the DipoleModel test was caused a wrong import. It was looking for
the module in the old package (propagator).

The EnviSat mission was added/updated so that it works with the new API. It also
was briefly tested and provides the same accuracy as under the old API.

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>