Created a general spacecraft object which provides the necessary methods for a spacecraft to be useful (builds the propagator, propagates, ...). Additionally two subclasses of spacecraft have been created. One to be used with the simulator providing the necessary methods for publishing the correct topics and communicating with other ROS nodes and one to be used with the planning module (also providing planning related methods)
Also the FileDataHandler file has been moved out to rospace_lib.misc at
it now is not only propagator specific anymore (but also used by the
planning module).

With this changes the QuickPropagator object is redundant and will be
deleted soon.

Some dirty hacks a still present in this commit and will be eliminated in
the near future.

The changes in this commit have been made based on the changes in the
planning repository (on the branch with the same name).

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>