Removed ugly hacks from Planning_Spacecraft and added docstrings.
The hacks have now been removed so that when the state of the
spacecraft is requested by the planning module it is extracted directly
from the propagator object. While this is computationally more
expensive it is a cleaner way of providing the necessary
spacecraft information (no multiple variables storing the same
things at different places how it was done before - which easily
results in errors/bugs).

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>