Scenarios are now called Missions. Each of those can now have multiple scenarios.
The so-called scenarios (which was used before) has now
been renamed to (the more logical) mission. Every mission now can
have multiple scenarios (initial conditions). This was done so
that the different scenarios from the monte carlo testing
of the relative navigation filter can be ported to new structure.

The difference now is that the simulator.launch file requires two
arguments which are a mission and its corresponding scenario.

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>