Quickly also fixed the bug concerning the imu frame broadcast.
The bug was a result of multiple spacecrafts having an imu
publishing to the same topic. The result was a flickering of
the frame as it's position is relative to the spacecraft frame
and due to the change of the reference frame the imu frame was
jumping between spacecrafts.

This was fixed by following the naming convention of sensors and
actuators in the cfg files. The sensor should all be called as
follows:
	<spacecraft>_s<number>_<sensor>
where:
	<spacecraft>: spacecraft name (e.g.: chaser)
	<number>: sensor number of spacecraft
	<sensor>: name of sensor (e.g.: imu)

The correct imu cfg for every spacecraft can now be found by
loading the sensors dictionary into the workspace and searching
for the sensor ending with the desired name (e.g.: imu).

This works like a charm.

Signed-off-by: Christian Lanegger <christian.lanegger@epfl.ch>